[Bronze III] Title: 다음수, Time: 132 ms, Memory: 14400 KB -BaekjoonHub

+++ 백준/Bronze/4880. 다음수/다음수.java
@@ -7,22 +7,20 @@ public class Main {
         
         while(true) {
             String[] input = br.readLine().split(" ");
-            double a = Double.parseDouble(input[0]);
-            double b = Double.parseDouble(input[1]);
-            double c = Double.parseDouble(input[2]);
+            int a = Integer.parseInt(input[0]);
+            int b = Integer.parseInt(input[1]);
+            int c = Integer.parseInt(input[2]);
             
             if(a == 0 && b == 0 && c == 0) break;
         
             if(c - b == b - a) {
-                sb.append("AP ").append((int) (c + (c - b)));
+                sb.append("AP ").append(c + (c - b));
             } else {
-                sb.append("GP ").append((int) (c * (c / b)));
+                sb.append("GP ").append(c * (c / b));
             }
             
             sb.append("\n");
         }
-        
-        
         
         System.out.print(sb.toString());
